+++ warp/__init__.pyi
@@ -260,7 +260,7 @@ from . import types as types
 from . import utils as utils
 from warp.config import DeterministicMode as DeterministicMode
 from warp._src.math import *
-from warp._src.marching_cubes import MarchingCubes as MarchingCubes
+from warp._src.geometry.marching_cubes import MarchingCubes as MarchingCubes
 from warp._src.context import RegisteredGLBuffer as RegisteredGLBuffer
 Length = TypeVar("Length", bound=int)
 Rows = TypeVar("Rows", bound=int)
